requirements.txt
pymongo
python-dotenv
rapidfuzz
pytz
openai>=1.0.0